Last working version

+++ src/structure.hpp
@@ -9,7 +9,7 @@
 struct symbols_t
 {
     float    value[GF];
-    uint16_t gf   [GF]; // to be removed !
+    uint8_t  gf   [GF]; // to be removed !
     bool is_freq;
 };
 //
